comments

+++ app/src/main/java/uk/ac/gre/mf9669w/m_expense/activities/MainActivity.java
@@ -2,11 +2,14 @@ package uk.ac.gre.mf9669w.m_expense.activities;
 
 import android.Manifest;
 import android.annotation.SuppressLint;
+import android.app.DatePickerDialog;
+import android.app.TimePickerDialog;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentSender;
 import android.content.pm.PackageManager;
 import android.location.LocationManager;
+import android.os.Build;
 import android.os.Bundle;
 import android.os.Looper;
 import android.view.Menu;
@@ -14,6 +17,7 @@ import android.view.MenuItem;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.ActivityCompat;
@@ -21,6 +25,7 @@ import androidx.core.content.ContextCompat;
 
 import com.google.android.gms.common.api.ApiException;
 import com.google.android.gms.common.api.ResolvableApiException;
+import com.google.android.gms.location.FusedLocationProviderClient;
 import com.google.android.gms.location.LocationCallback;
 import com.google.android.gms.location.LocationRequest;
 import com.google.android.gms.location.LocationResult;
@@ -32,6 +37,8 @@ import com.google.android.gms.tasks.Task;
 import com.google.android.material.snackbar.Snackbar;
 
 import java.util.Arrays;
+import java.util.Calendar;
+import java.util.Locale;
 import java.util.Objects;
 
 import retrofit2.Call;
@@ -48,25 +55,51 @@ import uk.ac.gre.mf9669w.m_expense.retrofit.RetrofitHelper;
 public class MainActivity extends AppCompatActivity {
     private static final int REQUEST_CHECK_SETTINGS = 100;
     private ActivityMainBinding binding;
-    private String nameOfPlace, destination, dateOfTrip, description;
+    private String nameOfPlace, destination, dateOfTrip, description, startTime, endTime;
     private boolean riskAssessment;
-    private double latitude, longitude;
     private LocationRequest locationRequest;
     private DBHelper dbHelper;
+    private FusedLocationProviderClient client;
 
 
+    @RequiresApi(api = Build.VERSION_CODES.N)
     @SuppressLint("MissingPermission")
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         binding = ActivityMainBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
+        client = LocationServices.getFusedLocationProviderClient(this);
         dbHelper = new DBHelper(this);
         Intent intent = getIntent();
         Trip trip = intent.getParcelableExtra("trip");
         if (trip == null) {
-            binding.saveDetailsBtn.setOnClickListener(v -> saveDetails());
-            checkPermissions();
+            binding.dateEt.setOnFocusChangeListener((v, hasFocus) -> {
+                if (v.getId() == R.id.date_et && hasFocus) {
+                    DatePickerDialog datePickerDialog = new DatePickerDialog(MainActivity.this);
+                    datePickerDialog.setOnDateSetListener((view, year, month, dayOfMonth) ->
+                            binding.dateEt.setText(String.format(Locale.getDefault(), "%d-%d-%d", dayOfMonth, month, year)));
+                    datePickerDialog.create();
+                    datePickerDialog.show();
+                }
+            });
+            binding.startAtEt.setOnFocusChangeListener((view, hasFocus) -> {
+                if (view.getId() == R.id.start_at_et && hasFocus) {
+                    TimePickerDialog timePickerDialog = new TimePickerDialog(this, (view12, hourOfDay, minute) ->
+                            binding.startAtEt.setText(String.format(Locale.getDefault(), "%d:%d", hourOfDay, minute)),
+                            Calendar.HOUR_OF_DAY, Calendar.MINUTE, true);
+                    timePickerDialog.show();
+                }
+            });
+            binding.endTimeEt.setOnFocusChangeListener((view, hasFocus) -> {
+                if (view.getId() == R.id.end_time_et && hasFocus) {
+                    TimePickerDialog timePickerDialog = new TimePickerDialog(this, (view13, hourOfDay, minute) ->
+                            binding.endTimeEt.setText(String.format(Locale.getDefault(), "%d:%d", hourOfDay, minute)),
+                            Calendar.HOUR_OF_DAY, Calendar.MINUTE, true);
+                    timePickerDialog.show();
+                }
+            });
+//            checkPermissions();
             if (isGPSEnabled()) {
                 getCoordinates();
             } else {
@@ -108,13 +141,25 @@ public class MainActivity extends AppCompatActivity {
                     }
                 });
             }
+            binding.saveDetailsBtn.setOnClickListener(v -> saveDetails());
         } else {
             binding.nameEt.setText(trip.getNameOfPlace());
             binding.destinationEt.setText(trip.getDestination());
             binding.descriptionEt.setText(trip.getDescription());
             binding.dateEt.setText(trip.getDateOfTrip());
-            binding.latitudeEt.setText(String.valueOf(trip.getLatitude()));
-            binding.longitudeEt.setText(String.valueOf(trip.getLongitude()));
+            binding.startAtEt.setText(String.valueOf(trip.getStartTime()));
+            binding.endTimeEt.setText(String.valueOf(trip.getEndTime()));
+            if (binding.dateEt.hasFocus()) {
+                binding.dateEt.setOnFocusChangeListener((v, hasFocus) -> {
+                    if (v.getId() == R.id.date_et && hasFocus) {
+                        DatePickerDialog datePickerDialog = new DatePickerDialog(MainActivity.this);
+                        datePickerDialog.setOnDateSetListener((view, year, month, dayOfMonth) ->
+                                binding.dateEt.setText(String.format(Locale.getDefault(), "%d-%d-%d", dayOfMonth, month, year)));
+                        datePickerDialog.create();
+                        datePickerDialog.show();
+                    }
+                });
+            }
             if (trip.isRiskAssessment()) {
                 binding.yes.setSelected(true);
             } else {
@@ -131,6 +176,8 @@ public class MainActivity extends AppCompatActivity {
         destination = Objects.requireNonNull(binding.destinationEt.getText()).toString().trim();
         dateOfTrip = Objects.requireNonNull(binding.dateEt.getText()).toString().trim();
         description = Objects.requireNonNull(binding.descriptionEt.getText()).toString().trim();
+        startTime = Objects.requireNonNull(binding.startAtEt.getText()).toString().trim();
+        endTime = Objects.requireNonNull(binding.endTimeEt.getText()).toString().trim();
         if (nameOfPlace.isEmpty()) {
             binding.nameEtLayout.setError("Cannot be empty");
         } else {
@@ -141,7 +188,7 @@ public class MainActivity extends AppCompatActivity {
                     binding.dateEtLayout.setError("Cannot be empty");
                 } else {
                     binding.radioGroup.setOnCheckedChangeListener((group, checkedId) -> riskAssessment = checkedId == binding.yes.getId());
-                    Trip trip = new Trip(tr.getId(), nameOfPlace, destination, dateOfTrip, description, riskAssessment, latitude, longitude);
+                    Trip trip = new Trip(tr.getId(), nameOfPlace, destination, dateOfTrip, description, riskAssessment, startTime, startTime);
                     boolean status = dbHelper.updateTrip(trip);
                     if (status)
                         Snackbar.make(binding.getRoot(), "Record updated successfully", Snackbar.LENGTH_LONG).setAnimationMode(Snackbar.ANIMATION_MODE_FADE).show();
@@ -155,6 +202,8 @@ public class MainActivity extends AppCompatActivity {
         destination = Objects.requireNonNull(binding.destinationEt.getText()).toString().trim();
         dateOfTrip = Objects.requireNonNull(binding.dateEt.getText()).toString().trim();
         description = Objects.requireNonNull(binding.descriptionEt.getText()).toString().trim();
+        startTime = Objects.requireNonNull(binding.startAtEt.getText()).toString().trim();
+        endTime = Objects.requireNonNull(binding.endTimeEt.getText()).toString().trim();
         if (nameOfPlace.isEmpty()) {
             binding.nameEtLayout.setError("Cannot be empty");
         } else {
@@ -165,7 +214,7 @@ public class MainActivity extends AppCompatActivity {
                     binding.dateEtLayout.setError("Cannot be empty");
                 } else {
                     binding.radioGroup.setOnCheckedChangeListener((group, checkedId) -> riskAssessment = checkedId == binding.yes.getId());
-                    Trip trip = new Trip(0, nameOfPlace, destination, dateOfTrip, description, riskAssessment, latitude, longitude);
+                    Trip trip = new Trip(0, nameOfPlace, destination, dateOfTrip, description, riskAssessment, startTime, endTime);
                     boolean status = dbHelper.insertTrip(trip);
                     if (status)
                         Snackbar.make(binding.getRoot(), "Record inserted successfully", Snackbar.LENGTH_LONG).setAnimationMode(Snackbar.ANIMATION_MODE_FADE).show();
@@ -219,6 +268,14 @@ public class MainActivity extends AppCompatActivity {
     }
 
     @SuppressLint("MissingPermission")
+    public void getLocation() {
+        client.getLastLocation().addOnSuccessListener(location -> {
+//            binding.latitudeEt.setText(String.valueOf(location.getLatitude()));
+//            binding.longitudeEt.setText(String.valueOf(location.getLongitude()));
+        });
+    }
+
+    @SuppressLint("MissingPermission")
     public void getCoordinates() {
         LocationServices.getFusedLocationProviderClient(this)
                 .requestLocationUpdates(locationRequest, new LocationCallback() {
@@ -229,10 +286,10 @@ public class MainActivity extends AppCompatActivity {
                                 .removeLocationUpdates(this);
                         if (locationResult.getLocations().size() > 0) {
                             int index = locationResult.getLocations().size() - 1;
-                            latitude = locationResult.getLocations().get(index).getLatitude();
-                            longitude = locationResult.getLocations().get(index).getLongitude();
-                            binding.latitudeEt.setText(String.valueOf(latitude));
-                            binding.longitudeEt.setText(String.valueOf(longitude));
+//                            latitude = locationResult.getLocations().get(index).getLatitude();
+//                            longitude = locationResult.getLocations().get(index).getLongitude();
+//                            binding.latitudeEt.setText(String.valueOf(latitude));
+//                            binding.longitudeEt.setText(String.valueOf(longitude));
                         }
                     }
                 }, Looper.getMainLooper());
@@ -258,9 +315,17 @@ public class MainActivity extends AppCompatActivity {
         if (item.getItemId() == R.id.view_trips) {
             startActivity(new Intent(this, TripsActivity.class));
         } else if (item.getItemId() == R.id.delete_all_data) {
-            boolean status = dbHelper.deleteAllData();
-            if (status)
-                Snackbar.make(binding.getRoot(), "Records deleted successfully", Snackbar.LENGTH_LONG).setAnimationMode(Snackbar.ANIMATION_MODE_FADE).show();
+            AlertDialog.Builder builder = new AlertDialog.Builder(this);
+            builder.setTitle("M-Expenses");
+            builder.setMessage("Are you sure you want to delete all data?");
+            builder.setPositiveButton("Yes", ((dialog, which) -> {
+                boolean status = dbHelper.deleteAllData();
+                if (status)
+                    Snackbar.make(binding.getRoot(), "Records deleted successfully", Snackbar.LENGTH_LONG).setAnimationMode(Snackbar.ANIMATION_MODE_FADE).show();
+            }));
+            AlertDialog alertDialog = builder.create();
+            alertDialog.show();
+
         }
         return true;
     }
